Networking Infrastructure (#4)
Add networking infra for sending beacons

For node, we send a regular post request. For IE11, we use a polyfill (same as the SDK). If beacon is natively supported, we use that instead. Also include a test site for testing manually and for tseting node

J=SLAP-1607
TEST=manual

Add test site and see that the beacon is sent on IE11, chrome, and on node