Update models for stronger typing (#5)
Rework the models so that Typescript is able to fully enforce the properties

Instead of having an AnalyticsEvent which takes an event name and a second param with takes the properties, if we pass a single object with a type property, typescript will be able to fully enforce the required properties. Therefore we can remove the AnalyticsEventData object and instead create data models for the entire event which also includes the event type.

I created a helper called EnumOrLiteral which allows the properties to accept either an Enum or the string literal associated with that Enum

I also added queryId because I found that the backend required that for most events

Add `SearchBarImpression` event as well

BaseData was also removed from the config because the data that needs to be sent with every event is experienceKey, experienceVersion, and businessId which are already sent on every event at they are included in the AnalyticsConfig. Furthermore, we don't want to allow people to send arbitrary data with the events, and there is no data that is shared across all analytics events that justify keeping the baseData option. It was mostly a relic from the SDK.

J=none
TEST=manual

Use the test site and test sending various types to the backend and ensure they are accepted