Support `pageDomain` query param (#35)
Allow users to specify the `pageDomain` query param as part of the `PagesAnalyticsConfig` in `providePagesAnalytics`. This support is needed to resolve an [issue](https://github.com/yext/pages/issues/229) in the PagesJS repo.

J=SLAP-2402
TEST=auto, manual

Add Jest tests and link the changes locally to the PagesJS repo and see that the `AnalyticsProvider` can be used to set the `pageDomain` through `providePagesAnalytics`.