Support the EU Region (#39)
Add support to the Search, Chat, and Pages analytics reporters for the EU

As part of this change, also add support for Sandbox to Search in order to make it consistent with Chat's support of Sandbox.

The yext-pixel.com domains have been updated to the new domains, and all of the code which constructs the URL endpoints is now consolidated in the provideEndpoints.ts file.

J=BACK-2262
TEST=auto, manual

Update existing tests and add new tests which confirm that the domains are correct when Search is configured for the EU or Sandbox.

Set the region and env in the test site for search, pages, and chat and confirm the requests are fired properly and the network responses are as expected. Check the EU db table for the search events and confirm they are present.