EventsSDK: Support Custom Actions in EventPayload
In the initial beta release we did not include custom actions as a part
of the Action type. This change adds support for an action field that
starts with 'c_' or 'C_'. This also updates some comments and
documentation.

J=FUS-5894
R=abenno, hyu
TEST=auto

Updated existing tests to use a custom action. Ran npm run test after
building and all pass.

Change-Id: I383636e369cb0fad5d69ce21affde36cfb9f27f4