EventsSdk: Update report return type
Remove errors from EventApiResponse. Instead we will report errors by
calling Promise.reject() which will print out the errors from the events
api. If it is successful, we return (but do not print) the result json
with the event id if called from fetch, and a boolean as a string if
called from post.

J=FUS-5895
R=mtian, abenno
TEST=manual

Made valid and invalid requests to the test site and ensured expected
behavior.

Change-Id: I1ecd6aa628c113b93c0ba7b10837133c9e53af5a