EventsSDK: Support Google Tag Manager
In order to support Google Tag Manager, we will export a new function
just specifically for sending Events API requests from it. This also
adds a convertTypesGTM function which is used to convert booleans and
numbers represented as strings to their correct types. This is unit
tested in convertTypes.test.ts.