EventsSDK: Revise Authorization in AnalyticsConfig (#124)
Previously we had to fields in AnalyticsConfig that we're optional,
`key` and `bearer`, where we throw an error if none or both are set.
This change updates the interface so it is more clear that only one is
required. There is now `authorizationType` which is either a string
`apiKey` or `Bearer` and `authorization` which is the token itself. We
need `authorizationType` so we can append the proper prefix of `KEY ` or
`Bearer ` to the final authorization property to avoid passing this work
onto the user. This also updates necessary test files to reflect this
change, and removes logic for ensuring that only one of key or bearer is
set.

[J=FUS-6201](https://yexttest.atlassian.net/browse/FUS-6201)
R=abenno, mtian, mkilpatrick
TEST=auto

---------

Co-authored-by: Ethan Jaffee <ejaffee@yext.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>