EventsSDK: make entityId optional for search's CtaEvent (#147)
In search, a click event fired on a Generative Direct Answer card might
not have an entity associated with it. Specifically, if the direct
answer text has a link, we do not know which entity that link comes
from, but we want to be able to report the link click event as a CTA
event.